tumor-segmentation/requirements.txt
tensorflow-cpu==2.10.0
Werkzeug==2.2.2
Flask==2.2.2
numpy==1.23.4
matplotlib==3.6.1
scikit-image==0.19.3
requests==2.28.1
Pillow==9.2.0
gunicorn==20.1.0
boto3==1.26.4